style(backend): clear all clippy warnings (stage 5)
Auto-fixes via `cargo clippy --fix` plus manual review:
- ToString impls -> Display (ResourceType/ResourceCategory/AuditStatus,
  UserRole, AuditAction); drop the now-redundant .to_string() in a log arg.
- min().max() -> clamp() for pagination/limit getters (behavior-preserving,
  bounds are constant with lo <= hi).
- Collapse nested if-let in api/oss.rs; derive Default on resource enums;
  and the batch of machine-applicable fixes (redundant closures, needless
  returns, etc.).
- AppState::new: #[allow(clippy::too_many_arguments)] with a note.
- Simplify the redundant compile-time type-check test in db/connection.rs
  (drop the wrapper fn and the meaningless assert!(true)).

`cargo clippy --all-targets` is now 0 warnings; 120 tests pass.

+++ backend/src/services/favorite_service.rs
@@ -261,7 +261,7 @@ impl FavoriteService {
                 .iter()
                 .max()
                 .copied()
-                .unwrap_or(0) as i32;
+                .unwrap_or(0);
 
                 FavoriteResourceItem {
                     id: row.id,
